requirements.txt
# Core async framework
aiohttp>=3.9.0
httpx>=0.27.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# AI Provider SDKs
anthropic>=0.40.0
openai>=1.58.0
google-generativeai>=0.8.0

# CLI & TUI
typer>=0.12.0
rich>=13.7.0
click>=8.1.0

# Data & Storage
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
redis>=5.0.0
chromadb>=0.5.0
msgpack>=1.0.0

# Document Processing
pypdf>=5.0.0
pdf2image>=1.17.0
pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
pdfplumber>=0.11.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation>=0.48b0
prometheus-client>=0.20.0
structlog>=24.1.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
tenacity>=8.2.0
watchfiles>=0.21.0
orjson>=3.9.0
pyyaml>=6.0.0
